Adding benchmark for sphere intersection.
